Changed default data format of multibeam data from floats to bytes. Fixed a bug in the multibeam reading program when the dispersion delays    are very large (greater than the length of one record).
git-svn-id: file://localhost/data1/git/presto_svn/trunk@118 5fac236c-a0d8-0310-9745-ef116914eda1

+++ src/prepfold.c
@@ -5,10 +5,10 @@
 /* Some function definitions */
 
 int (*readrec_ptr) (FILE * infile, float *data,
-		    long numpts, long numchan);
+		    int numpts, int numchan);
 int read_resid_rec(FILE * file, double *toa, double *obsf);
-int read_floats(FILE * file, float *data, long numpts, \
-		double *dispdelays, long numchan);
+int read_floats(FILE * file, float *data, int numpts, \
+		double *dispdelays, int numchan);
 
 /* The main program */
 
@@ -557,8 +557,8 @@ int main(int argc, char *argv[])
 }
 
     
-int read_floats(FILE * file, float *data, long numpts, \
-		double *dispdelays, long numchan)
+int read_floats(FILE * file, float *data, int numpts, \
+		double *dispdelays, int numchan)
 /* This routine reads a numpts records of numchan each from */
 /* the input file *file which contains normal floating      */
 /* point data.                                              */
